Optionally attempt to preserve line numbers in code output.
This uses line numbering data from the ast as captured by emit, which
currently means that only the introduction of a new calling form can
trigger a newline; we will want to allow symbols to do this too.

You have to pass {accurate=true} in options to compile or eval to get
this behavior.