Treat nil as a self-evaluating symbol in the AST instead of actual nil.
Given that Lua tables cannot store nil and lisp code is stored in data
structures, we either must cheat on tables and try to find a way to
make them store nil, or cheat on nil and find a way to store it in tables.

I believe the second approach is simpler since the symbol `nil`
naturally evaluates to nil; the only hiccup is that it must be
disallowed from being used as the name of a local. There are only two
places function parameters are introduced (destructuring and fn) so
catching those two is pretty straightforward.

I believe this would allow us to clean up a lot of AST-using code
since we could now trust Lua's built-in length operator to do the
right thing and stop tracking the .n field of parsed lists. Having to
ensure the correct value in the .n field was one surprising thing I
found quite error-prone when writing my own macros.