Revert 63c7c05 to fix if soundness
This Fennel code
```
(if (: (a) :a)
    1
    (: (b) :b)
    2
    ; else
    3)
```
compiles to this Lua
```
  local _0_ = a()
if _0_[("a")](_0_) then
  return 1
  local _1_ = b()
elseif _1_[("b")](_1_) then
  return 2
else
  return 3
end
```
which is unsound, as the temporary variable `_1_` that holds the result
of `(b)` is created under the previous branch as dead code.

A future solution would involve only generating elseifs if the code can
be guaranteed to not introduce any locals.