Catch typos that aren't in allowed globals list.
Note that this uses a mutable top-level local variable, because the
alternative discussed in #77 is very invasive.

This is off by default unless you pass allowedGlobals in the options
table when invoking the compiler. In the future it may be possible to
seed it with the default set of keys in _G, but that will require more
discussion.

Fixes #77.