Prettify table and comparator output
Tables themselves now are not wrapped in parentheses. Everything I've
ran passed, but if this does break code, tests should be added for those
situations and the parentheses should be able to be added for that
situation only, when necessary.

defineComparatorSpecial now uses parentheses like
defineArithmeticSpecial does, wrapping the whole expression in them as
opposed to the left and right hand sides. This works because any
expression that could cause precedence confusion (arithmetic or
comparator operators) will be wrapped in parentheses, and the output
looks more like hand-written Lua. Example:

- Fennel: `(< 1 2)`
  - Old: `(1) < (2)`
  - New: `(1 < 2)`
- Fennel: `(< 1 2 3)`
  - Old: `(1) < (2) and ((2) < (3))`
  - New: `((1 < 2) and (2 < 3))`
- Fennel: `(and (< 1 2) (* 3 5))`
  - Old: `((1) < (2) and (3 * 5))`
  - New: `((1 < 2) and (3 * 5))`