Implement saveLocals feature of the repl.
This implements a feature whereby locals in the repl are preserved
from one chunk to the next, which is currently one of the biggest
annoyances in the Fennel repl.

For every compilation, when the saveLocals isn't false, the repl will
compile the input, then in the penultimate line, it will inject a
snippet of code which loops over all the top-level locals in the given
chunk using the debug.getlocal function, and saves them off to a
table. After the code has been evaluated, all the values which have
been stored in this table get spliced into the current environment.

This more or less gives you a repl situation that acts "the way you'd
expect", allowing fn declarations from one input to remain around for
subsequent inputs.

The drawbacks are such:

* It introduces a ___replLocals___ global into the environment in
  order to have a place to store the values.

* Any local set in a chunk essentially becomes a global for the
  purposes of subsequent calls. This could have unintended consequences.

* It relies on assumptions about the string output of the compiler,
  namely that the last line is always a return, and that it's safe to
  inject code into the line right before the last line.