Allow destructuring in key/value tables, not just sequential.
This is actually a pretty small change; it seems like there's more to
update in the documentation than the size of the code itself. The main
thing is that we change from using ipairs to pairs when iterating over
the elements of a destructuring table.

The only gotchas are that we need a special check to only look for
rest args for sequential tables, and we need to do escaping on the
keys when emitting them if they are strings.