Make wrapEnv work with __pairs metamethod.
When an environment is provided to run code in, we have to wrap it
with a metatable in order to support mangling of globals. But because
the actual table doesn't have any keys in it, iterating thru _G to
list all the globals in the environment will not work unless the
metatable for _G includes a __pairs metamethod to delegate thru to the
provided environment for iteration just like it does for lookup/insertion.

There's probably a more efficient way to implement this, but iterating
over _G is not a common thing you do in tight loops; it is mostly only
used for discovery in the repl or in order to support tab completion
during development.