Work around openresty issue loading Fennel with _G metamethods.
In openresty, _G is an empty table with an __index metamethod. This
causes the globals checker to return an empty table, which causes the
compilation of built-in macros to fail. If we turn off the globals
checking for the built-in macros, it will still fail by default for
other calls, but the user can turn off globals checking in that context.