Don't error in fennelview when tables wrap sub-tables.
It's possible to have a table using metamethods to wrap all tables
inside it, for instance in order to enforce read-only behavior. When
this happens, it can return a new table every time that key is
looked up. This causes the "appearances" tracking (used in cycle
detection) to fail to track a given table.

This change makes cycle detection fail gracefully in that case instead
of erroring out. If you introduce metamethod wrappers that contain
cycles, sorry--it's on you to sort that out yourself. We can't catch it.