Require autogensym for identifiers inside backticked macros
When using backtick, in order to avoid symbol capture, bare symbols
are not allowed to be used as identifiers. Instead, gensyms must be
used. In order to make this more palatable, we support "auto-gensym"
by appending # to the end of your locals; within a single scope this
will create a gensym beginning with your local name.

This is similar to the method Clojure uses to prevent accidental
symbol capture; however Clojure does this by auto-qualifying all
symbols inside a backtick with the current namespace. Since we don't
have namespaces, we simply set a "quoted" flag on the symbol.

This also reverts the change which prevented # from being a valid
symbol constituent character.