Enable lua style method calls
(foo:bar baz) is now syntax sugar for (: foo :bar baz). reference.md
modified to list this form of method calls first, with the
(: foo :bar baz) style as a backup. : is now treated as a multisym
separator in symbols when not in the initial position, only one is
allowed per multisym, and it must be the last multisym separator.