Fix bad code gen.
This was due to disallowed ambiguity in the Lua grammar. If a statement
begins with an open paren, it could be interpretted as a function call
or as a grouping.  Lua simply rejects this syntax in certain cases. This
fix prefixes problematic statements with a semicolon, which is a
statement separator to resolve ambiguity.