Improve documentation.
The main thing here is a more thorough explanation of the io and table
modules in the Lua primer, but also an attempt to clarify the
relationship between the different pieces of documentation and where
you can find what you're looking for.

Also adds a mention of `while` to the tutorial.