Work on #218
This change prevents aliasing between new locals declared
with declareLocal. New locals, instead of shadowing locals with the same
name, instead will have a new, unique name. In order to make this work,
we need to be able to postpone the application of manglings on symbols
after calling declare local.

Consider:
(local a 1)
(local a (+ a 1))

Without postponing the application of mangling, the resulting Lua would
be:

local a = 1
local a0 = a0 + 1

While instead, we want

local a = 1
local a0 = a + 1

The new function applyManglings (used internally) and an extra argument
to declareLocal enable this functionality. These functions are needed to
be able to call declareLocal on the lvalue _before_ calling compile1 on
the rvalue.

With all of these changes, the aliasing problems as presented in #218
should go away. However, this may have other implications.