Disallow locals aliasing globals
If a global reference is used inside the definition
of a local with the same name, the global is aliased by the local.
Previous attempts to get around this do not work generally, but we can
detect when such a situation occurs and emit a compiler error. This
'fix' could possibly be better addressed with a VERY large compiler
refactor, but such code is usually bad style anyways.