Use metatables for storing source data on table literals.
For most AST nodes we store source data in the node itself, but table
literals are themselves, so if we stuff new fields in them, they'll
show up in the output! No good. Using the metatable is a nice place to
hide them away so they are available but don't get in the way.

Before:

    Compile error in test/bad/unable-to-bind.fnl:1
      unable to bind number 1

    (let [1 1] 1)
    ^^^^^^^^^^^^^

After:

    Compile error in test/bad/unable-to-bind.fnl:1
      unable to bind number 1

    (let [1 1] 1)
         ^^^^^