Address code review suggestions.
* Add macros to completion targets.
* Add docstring to macrodebug.
* Fix fennelview module typo.
* Clarify macrodebug only works on top-level macro calls.
* Pass macroCurrentScope thru in macroexpand function.

That final point makes it so that macroexpand is not usable from
general eval-compiler calls but only from within eval-compiler. This
is because we don't have a way to pass in the correct scope arg for
non-top-level calls, so it's better to limit its applicability rather
than imply that we can do proper arbitrary scoping here.