Make fennel launcher script self-hosting.
This is a big change!

This moves the existing `fennel` script to `launcher.lua` and
introduces a new `launcher.fnl` script which is a pretty straight
port of the old version, except since it's in Fennel, we can use the
--require-as-include option in order to create a standalone script
that works as a single file.

This will make installation of releases much more straightforward
since now there is a single file.

There are two downsides:

1. There is risk that the port from Lua to Fennel introduced bugs.
2. Now when you do a fresh checkout, you must run `make` before you
   have a `fennel` script you can use. Every change you make to
   fennel.lua, fennelview.fnl, or fennelfriend.fnl will not get picked
   up until you run `make` again.

We could check a compiled `fennel` into the repository to help with
the first half of 2, but not the second part.

Fixes #248.