Add --compile-splice.
This allows binary builds of Fennel to "splice" themselves: to compile
new standalone Fennel programs based on themselves without any need
for gcc, liblua, or the Lua dev headers.

This is accomplished by making a copy of the compiled binary and
appending the Lua code to the end of it, followed by a comment
indicating the offset in the file at which the Lua code begins.

When the binary launches, it looks at the end of itself to see if such
a comment exists. If it does, it tries to load it; otherwise it
continues as per normally, just acting as the Fennel executable.