make searchModule respect package.config
For cross-platform compat, package.config contains info about what to
use as a dir separator, a delimiter in the package.path list, what to
use as a placeholder, etc. This change uses the first three values in
package.path in searchModule, which should allow "\" in fennel.path to
work on Windows.