Move straggler functions into pseudo-modules.
These don't always fit nicely, but they're better than having them
free-floating.

In particular all the "root"-related stuff goes into utils, but only
because utils is available everywhere.

dofile has to be put into the compiler module, not because it
logically makes sense, but because that's the module that's in scope
where dofile is needed. This one is particularly bad because it's a
circular dependency.

There is one other circular dependency; the `fennel` module must be
available in compiler scope, but it hasn't been defined yet when
makeCompilerEnv when is defined.