added 'coverage' target to Makefile using luacov
To get more accurate results, we'll likely need to tweak the settings a
bit in .luacov, but at least we now have a basic working setting for
generating coverage reports.

Not yet adding this to `make ci` - we can always do so once we play with
it more , and use a more readable coverage reporter such as
luacov-console, luacov-cobertura, luacov-coveralls, etc.