Move "outer" fennel module to being written in Fennel.
Each former pseudo-module (utils, parser, compiler, and specials) gets
moved to a real module, but still written in Lua.

The only changes to utils.lua, parser.lua, compiler.lua, and
specials.lua are to move them to their own files, make them no longer
buried inside an IIFE, and to move requires to the top of the file for
the modules they depend upon.

The repl gets moved from an inline string into a proper file.

The fennelview.fnl, fennelfriend.fnl, and fennelbinary.fnl files will
get moved into the src/ directory, but haven't been moved yet.

The old implementation is now in old/fennel.lua and is used to
bootstrap the new one. The fennel.lua file is now a generated file.