Prevent early evaluation in method calls and comparators.
Previously in certain situations we would use the internal compiler
helper function `once` in order to ensure a value only got evaluated
once, to prevent double-evaluation in method calls and
comparators.

Unfortunately this had a bug where in an effort to prevent
double-evaluation, it would evaluate it exactly once, rather than at
most once, because it would emit the expression evaluation and bind it
to a local in the parent context and then use the local in multiple
places.

This would cause problems because there are cases where it should not
be evaluated at all:

    (and false (: (foobar) :method?))

    (and false (< (a) (b) (c)))

Both these cases are special because they cannot compile into the
"lua-native" equivalent construct; in the first case because the
method name has a question mark in it, and in the second case because
the < operator has three args instead of two.

We fix this by compiling to an IIFE and using its arguments to refer
to values twice without danger of double-evaluation:

    (function(tgt, m, ...) return tgt[m](tgt, ...) end)(%s, %s)

Of course, this introduces some overhead, but it is the only safe way
to compile the code. In both cases we check to see if the lua-native
compilation approach is possible, and if it is we use it, so the
overhead will only be incurred when absolutely necessary.

Fixes #323