This patch adds support for sandboxing the compiler.
Previously there was no way to safely accept arbitrary
input since sandboxing would only apply to the compiled code, not to
macros and eval-compiler.

However, suddenly disallowing access to globals in compiler scope
would be a breaking change, so instead what we're doing is allowing
unsafe things thru but emitting a warning. In the future once we can
make a breaking change, (maybe for 1.0?) that warning will be changed
to an error.

The most awkward part about this change is its effect on require. We
want to allow modules to be loaded, but using a separate
compiler-scoped replacement for package.loaded which only allows
modules which have been loaded with the limited environment.

We have to allow `require` because of a quirk in how we compile
metadata; the compiled code requires the fennel module and sets the
docstring on the metadata field of it. So we have to allow requiring
the Fennel module. In the default compiler-env we replace require with
a new version which lets thru the module as needed for docstrings.

The sandboxing can be disabled by launching Fennel with the
flag --no-compiler-sandbox or (if you're using the compiler API) by
passing in :compiler-env _G in the options table.