Add plugin system, take 2.
This is roughly the same as the previous take on the plugin system,
which was based on the pre-self-hosted compiler.

Each plugin is a module which contains functions for various extension
points within the compiler. Each field in the module is named after an
extension point and should be a function that takes AST arguments and
a scope table. Current supported extension points are:

* symbol-to-expression
* call
* destructure
* do
* fn

These were chosen because they're the minimum needed to implement the
linting example, not by looking thru the whole compiler to find an
exhaustive list of good extension points. So we are likely to be
missing some good candidates for extension; we should be open to
adding new ones.

The included linter.fnl file shows an example of how the system works.

However, the problem here is that we are exposing compiler internals,
and that means that when we change the internals, we are likely to
break plugins when we refactor the compiler. I don't really know what
to do about that, and it seems inevitable. So we should probably just
try to set expectations that plugins will tend to be only compatible
with a limited set of compiler versions.

One problem is that the plugins need to be loaded in compiler scope,
which means we need a way to indicate in fennel.dofile that we
want to load the file in question in the compiler scope. In this patch
you can pass :env :_COMPILER as a special flag which causes the :env
to get replaced by a new compiler-scope environment. I don't love
this; I think we could come up with a better way to expose it.

We still need more documentation for this in api.md before the release.