first pass on fixing compiler-env propagation
To make macros imported by `import-macros` respect the use of
`--no-compiler-sandbox`, I added `utils` to the compiler-env for the
built-in macros so `import-macros` can use it when compiling the macro
module.

It works when tried manually from the CLI, however the new test does not
yet pass for reasons unknown.