Add comma-commands to repl.
The commands in the repl are "metacommands" in that they live
"outside" the environment of the repl, invoked by prefixing them with a
comma. This allows us to define commands in a way that we don't have
to worry about conflicts with user code, since the comma prefix is
normally only valid within the context of a quoted form.

The interface is inspired by a similar construct in SLIME's repl:

https://common-lisp.net/project/slime/doc/html/Shortcuts.html

Currently we have:

* exit
* reset
* help
* reload

Of these, the last is the most interesting; we take the basic approach
from lume.hotswap which allows retroactive reloading so that code
which already has a copy of the module table in question will have
access to the new updated fields. It's also the only command to take
an argument.