Add symmeta to declared locals later to work around shadowing problem
When you shadow an existing local, the right-hand side is compiled
after the local is declared, not before. The emitted code is still
correct, but the usage-tracking is incorrect.

The problem is that there is a circular dependency between the right
hand side and the left hand side. The RHS must know the names of the
locals it's getting bound to (see opts.target) but it also must be
compiled without those locals really being in scope yet, because those
locals aren't really "active" until after that binding has been
established.

This basically only matters for linting. The following code compiles
correctly, but fails to lint without this patch.

    (fn [abc] (let [abc abc] abc))

Without this patch the third abc gets treated as a reference to the
let-bound abc, not the function parameter.

Deferring the creation of the symmeta table allows the right-hand side
to compile without triggering the bug as far as pertains to usage tracking.

Fixes https://todo.sr.ht/~technomancy/fennel/12