Don't require --metadata for searcher in the repl
The --help text says that metadata is enabled by default in the repl,
but this is currently only true of code that's entered directly;
metadata on loaded modules is discarded.

This changes it so that the repl enables searcher metadata unless it's
specifically disabled.

But in order for that to work, it has to change the default searcher
function so that it keeps a copy of the options table and copies all
those options every time a file is found to load. So now changes to
the make-searcher options table are respected going forward rather
than needing the options to be decided when make-searcher is called.

This is a clear improvement from the perspective of the repl, but it
might lead to surprising behavior if people don't expect make-searcher
to respect changes to the table.