Add hints to internal destructuring variables
This patch adds a single option to destructure function, symtype, which
lets the compiler add the name of the syntax form performing
destructuring to the Lua output of the destructuring. Variables
generated look like _let_0_. The symbols generated for the left-hand
side of destructures are unchanged.

This patch has no effects on behaviour but should slightly clear up
reading and debugging compiled code.