use pretty-printed function arglist in fnl/arglist metadata
This change adds two new predicates - string? and keyword? and a new
deep-tostring function, which transforms literal tables into single
line string representation.  I've modified set-fn-metadata to use this
new function, and updated tests.  This is not as general as
fennelview, but given that we only need to print literal tables when
creating documentation string, I think that this is enough.