In assert-compile, don't assume ast argument is a table.
It shouldn't be a problem if a macro calls assert-compile with an ast
which is a non-table value, since the caller of the macro could
provide just about anything as the argument in question. Rather than
making every macro check for a table before calling it in
assert-compile, we should do the check once inside assert-compile. If
a non-table is provided, fall back to "unknown" values.

Note that this needs to be done twice, once for "friendly" errors and
once for the simpler fallback "unfriendly" assertions.

Fixes https://todo.sr.ht/~technomancy/fennel/38