Allow parsing of comments as an opt-in feature.
Normally we discard comments during parsing because it would be
confusing in macros to have to deal with nodes that have no semantic
value. But sometimes when you want to use the parser standalone (for
instance, when formatting code) you want the comments included.

Separated out the parser tests into their own file.