Ensure stablepairs never attempts to index nil.
I ran into an issue when integrating with lite. Lite implements a
"strict mode" by assigning a metatable to _G that throws an error
when an undefined global is accessed or set.
https://github.com/rxi/lite/blob/master/data/core/strict.lua

The implementation of stablepairs in 0.8.0 assumes that indexing
(. tbl nil) will harmlessly return nil. This is not true in the
presence of metatables. In lite's case, `current-global-names` ends
up calling it on _G; this causes _G[nil] to be accessed and an error
to be thrown.

Jeremy