ensure that __fennelview has higher priority than :prefer-colon?
When __fennelview calls view function with last argument set to true
the string is printed as colon string, when possible.  When the last
argument is false, string is always printed in double quotes.  When
the last argument is omitted, value of :prefer-colon? key is used.

I've also used forward declaration for internal pp function instead of
defining it in a table, and fixed docstring, as per 0e8efe6 the
docstring mentions :colon-strings? instead of :prefer-colon?  option.