fix ?. behavior introduced in 18c6fcf and fix tests not being ran
The change introduced in 18c6fcf made ?. behave the same as .
This was not caught by the tests, because tests were not ran at all,
because "expected" result was nil, and assigning nil to table key
effectively removes that key from the table, thus making tests
completely ignored. I've noticed this and fixed in some other tests in
core.fnl as well.