Store comments inside k/v tables differently when parsing.
These can't be kept inside the table itself because they have to have
an even number of elements, and because they're un-ordered.

This also introduces the `keys` table on the metatable, which allows
you to reconstruct the ordering of a table even after it's gone thru
the parser and has resulted in an un-ordered AST node.

Still need some documentation around the AST format.