Only emit disambiguating semicolon when it's needed.
Tragically, Lua is not a lisp.

As seen in https://todo.sr.ht/~technomancy/fennel/52 the semicolon we
used to disambiguate grouping vs function calls is allowed anywhere in
PUC Lua, but not in LuaJIT. LuaJIT refuses to allow the semicolon if
it's not actually separating statements, so we check to make sure that
the code being emitted is not the first thing in the chunk.