Fix correlation being broken with macros
Before, a program like (when true 4) would input sixty-odd lines of
space when compiled with --correlate. This is because the invocation
of the when macro caused the AST to include a node with a line number
of the code *in macro.fnl*. The compiler already checked for the file
name of the syntax tree and the currently-compiled file being equal,
but both checks were broken and always returned nil (without an error),
so the correlation always happened.

This patch might break something, since it messes with some code that
has been silently failing for a bit.