?. macro: further improve the emitted lua
The `(-?> t (. a) (. b) (. c))` version was an improvement over the one
before it, since instsead of deep-nesting a closure for each key, it
instead nested if-else-done blocks, but since each parent is only
accessed once, it can be further improved upon by reassigning to a
single symbol.

This, ironically, goes against the general code philosophy seen in
fennel, but results in the cleanest output, and macro hygiene will
prevent the downsides we'd otherwise be concerned with.

Compare the two outputs for the following:

```fennel
(?. {:a {:b {:c {:d :HEY}}}}
    :a :b :c :d)
```

```lua
-- new impl
local t_0_0 = {a = {b = {c = {d = "HEY"}}}}
if (nil ~= t_0_0) then
  t_0_0 = (t_0_0).a
end
if (nil ~= t_0_0) then
  t_0_0 = (t_0_0).b
end
if (nil ~= t_0_0) then
  t_0_0 = (t_0_0).c
end
if (nil ~= t_0_0) then
  t_0_0 = (t_0_0).d
end
return t_0_0

-- previous version that expands to (-?> t (. a) (. b) (. c))
local _0_0 = {a = {b = {c = {d = "HEY"}}}}
if _0_0 then
  local _1_0 = (_0_0).a
  if _1_0 then
    local _2_0 = (_1_0).b
    if _2_0 then
      local _3_0 = (_2_0).c
      if _3_0 then
        return (_3_0).d
      else
        return _3_0
      end
    else
      return _2_0
    end
  else
    return _1_0
  end
else
  return _0_0
end
```