compiler: allow strings at call position
Strings can possibly have a __call metamethod, and it's actually
possible to call strings in Lua with the following notation:
`("str")(...)`.  This patch adds a check if AST element at call
position is a string, and if it is, it is wrapped in parentheses to
ensure correct Lua syntax.

I wonder if we can wrap every literal into parentheses, as
something like this is valid lua:

local function foo(bar) return bar end
(foo)(42)

Also if __call metamethod for strings is not restored in the test, the
test-empty-values test breaks:

    ./test/luaunit.lua:2893: Expected an error when calling function
    but no error generated