Fix destructuring kv tables inside values
When symmeta is added in destructure-values, we have to check that the
value is in fact a sym (not a table). Any names declared in the table
destructure have their symmeta added in the recursive call to
destructure1.

This manifests as a compiler error:

    (let [({:x 10}) {:x 10}] x) => Error: table index is nil

... from setting a table key to (utils.deref {:x 10}), which is nil.

Though note that sequential table destructuring accidentally worked,
since utils.deref is just (. sym 1), or in this case, `x`:

    (let [([x]) [10]] x) => 10

This appears to be a regression from b43fe48.