Fix a completion/allpairs error on dynamic __index
utils.allpairs checks for, and iterates, any __index metamethods on
the target table; this leads to an error when attempting repl completion
on tables whose __index is dynamic (as is the case for some lua
modules).

This fix simply skips any dynamic __index, since those can't be iterated
anyway without also having a custom __pairs present.