Announce on libera chat when there has been a failure in CI.
This uses netcat to connect to the IRC server and send some raw
commands when we detect that a build failure has happened. We include
the output of `git log --oneline -n 1` in the message.

The announcement will only be sent when the `IRC_CHANNEL` env var is
set; currently that is set in `.build.yml` and nowhere else.

Also remove cloc from the CI runs; it doesn't really add anything for
every run and it has more dependencies than I would like.