Reset line/byte after each repl chunk.
Prior to this we were seeing in the repl that byte offsets and line
numbers in the AST were continuing to increase with every new line in
the repl, but the source being counted against was just the single
input. This caused error pinpointing to grow increasingly inaccurate
as more lines were entered.

See https://todo.sr.ht/~technomancy/fennel/74

Also fix a bug where fennel.friend miscalculated line offsets when
reading source from a string rather than from a file; previously this
was obscured by the above bug.