Check binding sequence in let.
We don't literally check for utils.sequence? because that checks for a
metatable which is set by the parser, but it's common in macros to
construct a sequence table without a metatable. This means that
certain kinds of nonsense is still allowed, but there is no way to
detect it without also blocking valid macros.

    (let {1 a 2 "hello"} (print a)) ; valid

Fixes https://todo.sr.ht/~technomancy/fennel/87