Treat else branches in if as nil, not zero values.
Previously when you had an even number of arguments in an if, there was
no else branch emitted, which means that if none of the conditions were
truthy, zero values were returned.

In Fennel we can't avoid some cases where zero values are distinct
from nil, simply because we must be compatible with Lua, and Lua makes
such a distinction. But there are some cases where we can avoid
zero-values, and this is one of them. If you really want zero values,
you can put (values) in your else clause.