Swap out locals-saving implementation with scope-aware version.
This was previously implemented by taking the compiled Lua output and
wrapping it in some other Lua strings. However, this meant that the
compiler was unaware of the locals which were getting spliced in,
which led to unpredictable behavior.

Now instead we wrap the parsed input form with another form which
serves the same purpose, but the entire form can be compiled as a
single unit, so the results are much more consistent. Plus we're not
splicing string together any more, which is gross!

Previously because we were splicing forms based on Lua lines, our
method only worked when there were multiple lines of Lua because we
put our locals-saving code before the last line. Now dealing with
forms we don't have that problem.

See https://todo.sr.ht/~technomancy/fennel/85#event-97190