Compile Fennel to bytecode during --compile-binary.
This should hypothetically produce binaries that are smaller and
faster to load; however that is not the case. The bytecode produced
actually has what appears to be the entire text source code embedded
inside it, so it is larger than skipping bytecode entirely. It appears
to have no effect on loading speed either, measured based on the time
taken to load Fennel itself. Perhaps other programs would yield
different results.

See https://p.hagelb.org/fennel.out for an example of the bytecode.

Further investigation is required.