Loosen the global/local shadowing rules.
That rule was put in place specifically to work around a bug in this code:

  (let [pairs #(pairs $)] pairs)

This code looks like it should be able to refer to pairs inside the
function, but it will in fact compile to an infinite recursion because
the function gets defined in such a way that its own definition
shadows the global it's trying to refer to.

The problem is that the code which tried to detect that case also
triggered on a number of completely harmless cases:

    (print (type :x))
    (let [type :string]
      (print type))

That would fail, whereas the same code without the top-level reference
to the global would be fine.

This code loosens the restriction so it only checks one level of
scope, not the entire chain.