Fix a bug where and/or didn't properly short-circuit.
The problem is that nval=1 was telling it that it could do the `do`
above without an IIFE, but without nval=1 we got multivals spliced
into the middle of arglists.

The solution is to compile them all the same (with nval=nil) and then
discard all but the first value for everything but the final arg.