Accept two body arguments in collect for separate key and value.
Previously the body of collect had to be a single expression which
compiled to multiple values; usually (values k v).

Now you can supply the key and the value as separate arguments and it
will turn it into (values k v) for you.

The old behavior is preserved for backwards compatibility.