Put compiler env values in provided env table instead of metatable.
This simplifies things greatly because we no longer need metatables
for compiler env sandboxing.

However, it has one downside; now we modify the table you pass in,
which could be surprising. The previous behavior *allows* the compiler
code to modify the provided env, but it doesn't modify it by default;
the env table only changes if the macro code changes it. (This is used
in many places to convey values out of compiler scope in the test
suite.)

Because we suggest passing in _G as the :compiler-env value when you
want sandboxing disabled, and modifying _G is a really bad idea, when
we detect _G is used as the compiler sandbox, we instead create a copy
of _G and use that as the env. In this case, values cannot be conveyed
out of the compiler env;  copy of _G is needed for that, which we do
in the tests.