Add match-> macro for chaining possibly-failing steps.
This is intended for use when chaining together several steps which
could fail, typically by returning nil+msg. The chain continues when
there is a match, but upon the first mismatch, the entire form will
evaluate to the mismatched value (nil+error).

This allows you to avoid tedious chains like how Lua often has:

    local ok, val = do_whatever()
    if ok ~= nil then ... end

Example:

    (match-> (get-file-name-for user)
      filename (io.open filename :w)
      file (file:write data)
      file (notify user))

The main downside of this implementation is that since we don't know
statically how many values each step will need, we are forced to place
each step inside a table and unpack it again when needed, creating
some performance overhead. If we knew that no step would ever have
more than just the two values of nil+error, we could have a simpler
and faster macro, but I believe it's more valuable to have the
flexibility of working with any number of values.

See https://todo.sr.ht/~technomancy/fennel/104