bugfix: PUC Lua 5.1-compatible escapes
Since compilation with PUC 5.1 used to work, I tried bisecting and
found that issues first appeared with commit 89296445

```
guix build fennel --with-commit=89296445 --with-input=lua=lua@5.1
```

It turns out that Lua 5.1 does not support hexadecimal escapes.
Works in LuaJIT though, because it's fancy like that. (:

I've replaced the handful of hexidecimal escapes with their decimal
counterparts.

Fixes: https://todo.sr.ht/~technomancy/fennel/108
Fixes: https://github.com/bakpakin/Fennel/issues/386