add missing utf8 library to OK'd compiler globals
There shouldn't be any particular need to sandbox it like `os` or `io` -
however, it's only available in Lua 5.3 and Lua 5.4.

Also took the opportunity to reimplement `utils.copy` in terms of
icollect. We could *probably* just remove it entirely, but it's somewhat
nice to have a function version around, and I'm not sure we wouldn't
end up writing more code we'd save by dropping the util, given how
short it is.