Fix a bug where disabling the compiler sandbox broke module require
When the compiler sandbox is disabled, previously requiring modules
would just use regular require, which loaded modules normally with
runtime scope. The propagation of compiler scope (a macro module
requiring another macro module and getting it treated as a macro
module) only happened within the compiler sandbox.

This change fixes that so even when the sandbox is disabled, compiler
scope still propagates.