Make varg no longer be a singleton.
This allows us to use error pinpointing on it like a sym.

Unfortunately we were exposing the singleton in compiler scope as
_VARG so if anyone was relying on the fact that it was a singleton, it
could break that code. But we have had the varg? predicate in compiler
scope for a very long time which is pretty clearly the right way to
check that.